Rename DB interface to 'Storage'
