+++ cli/commands/metric/metric_test.go
@@ -27,6 +27,7 @@ package metric
 
 import (
 	"bytes"
+	"clouditor.io/clouditor/persistence/gorm"
 	"fmt"
 	"io/ioutil"
 	"net"
@@ -51,7 +52,7 @@ var (
 	server              *grpc.Server
 	orchestratorService *service_orchestrator.Service
 	authService         *service_auth.Service
-	db                  persistence.IsDatabase
+	db                  persistence.Storage
 )
 
 func TestMain(m *testing.M) {
@@ -65,7 +66,7 @@ func TestMain(m *testing.M) {
 		panic(err)
 	}
 
-	db = new(persistence.GormX)
+	db = new(gorm.GormX)
 	err = db.Init(true, "", 0)
 	if err != nil {
 		panic(err)
