+++ cli/commands/service/discovery/discovery_test.go
@@ -27,6 +27,7 @@ package discovery
 
 import (
 	"bytes"
+	"clouditor.io/clouditor/persistence/gorm"
 	"fmt"
 	"io/ioutil"
 	"net"
@@ -53,7 +54,7 @@ var (
 	server           *grpc.Server
 	authServices     *service_auth.Service
 	discoveryService *service_discovery.Service
-	db               persistence.IsDatabase
+	db               persistence.Storage
 )
 
 func TestMain(m *testing.M) {
@@ -67,7 +68,7 @@ func TestMain(m *testing.M) {
 		panic(err)
 	}
 
-	db = new(persistence.GormX)
+	db = new(gorm.GormX)
 	err = db.Init(true, "", 0)
 	if err != nil {
 		panic(err)
