+++ persistence/persistence.go
@@ -31,10 +31,10 @@ var (
 	ErrRecordNotFound = errors.New("record not in the database")
 )
 
-// IsDatabase comprises a database interface
+// Storage comprises a database interface
 // TODO(all): I think with generics we could get sth. like `[T]Get(r T, id string) T` which is maybe more intuitive
 // then changing the parameter? Maybe its only my impression
-type IsDatabase interface {
+type Storage interface {
 	// Init initializes the database
 	// TODO(lebogg): Use NewService Pattern (maybe not here, though)
 	Init(inMemory bool, host string, port int16) error
